+++ src/components/Character/Character.tsx
