+++ src/components/Character/Characters.styles.tsx
