+++ src/components/Character/index.ts
@@ -1,0 +1,3 @@
+export { default as Character } from "./Character";
+export { default as Characters } from "./Characters";
+export { default as CharacterDetail } from "./CharacterDetail";
