--- src/components/Footer.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-
-import styled from "styled-components";
-
-export const Footer1 = styled.footer`
-  text-align: center;
-  opacity: 0.7;
-  margin-top: 40px;
-`;
-
-function Footer() {
-  return (
-    <>
-      <Footer1>© 2023 Shionpark</Footer1>
-    </>
-  );
-}
-
-export default Footer;
