+++ src/components/Footer/Footer.styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const FooterInner = styled.footer`
+  text-align: center;
+  opacity: 0.7;
+  margin-top: 40px;
+`;
