+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { FooterInner } from "./Footer.styles";
+
+function Footer() {
+  return (
+    <>
+      <FooterInner>© 2023 Shionpark</FooterInner>
+    </>
+  );
+}
+
+export default Footer;
