+++ src/components/Header/Header.styles.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Logo = styled.img`
+  width: 130px;
+  height: 50px;
+  margin: 0 auto;
+  display: block;
+`;
