+++ src/components/index.ts
@@ -1,5 +1,3 @@
-export { default as Character } from "./Character";
-export { default as Characters } from "./Characters";
-export { default as CharacterDetail } from "./CharacterDetail";
-export { default as Footer } from "./Footer";
-export { default as Header } from "./Header";
+export * from "./Character";
+export * from "./Footer";
+export * from "./Header";
