+++ src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Characters from "../components/Characters";
+import { Characters } from "@/components";
 import styled from "styled-components";
 
 const Title = styled.h1`
